+++ HISTORY.md
@@ -1,6 +1,7 @@
 unreleased
 ==========
 
+  * perf: hoist regex declaration
   * perf: use regex to extract mime
 
 2.0.2 / 2015-01-31